Модуляция + добавил функцию Генератор для Object.id & Object.url
